requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
requests>=2.25.0
fpdf2>=2.5.0
inference-sdk>=0.0.1
qrcode[pil]>=7.0.0
python-dotenv>=0.19.0


